Add exclusion rules to filter unwanted events (#259)
* Add exclusion rules to filter unwanted events

* Add json logic error handling

* Fix format/imports with goimport

* Rework tests to cater for out of order events

Looks like the ordering for events received from the channel is
non-deterministic. I had changed the tests in this PR to expect events
in an exact order, which passed reliably on my local machine but failed
when run by the GitHub Actions workflow. In this commit I've reverted
the existing big picture test (the one with no exclusion rules) to not
check the payload, and modified the new test (including an exclusion
rule) to:

a) wait 1 second for all events to be received on the channel
b) verify that the excluded event is not received regardless of order
c) verify that the expected number of events is received

* Fix format